Complete Refactor of the Plex Library to handle net.Conn specifically
After going through the process of building the `plex` library to
handle the io.Reader/io.Writer/io.ReadWriter types I realized once I
began buidling the more operational unit tests that I had abstracted
the library too far.

The intention with `plex` was always to handle multiplexing of net.Conn
intances since it will layer on top of the Atomizer communication
system. Because net.Conn implemented the io.ReadWriteCloser type I
thought that abstracting it out that far would work. It lead to a great
deal of code that was not directly related to the net.Conn work and led
to even more headache in the form of weird races in handling the IO types
which were never the sole purpose.

There were also design decisions that did not make sense for an IO type
like auto-scaling, self-healing or multi-stream multiplexing which are
much more intuitive when viewed as a library for handling net.Conn.

All of this led to a much more fine tuned library with code that is
more easily reasoned about.